+++ src/cli-helper.js
@@ -7,6 +7,7 @@ const path = require('path');
 const { spawn } = require('child_process');
 const checkForUpdate = require('update-check');
 const isDocker = require('is-docker');
+const { weiAmountSchema } = require('./validator');
 const { storageProviders } = require('./params-utils');
 const packageJSON = require('../package.json');
 
@@ -28,10 +29,11 @@ const info = {
   showing: obj => `Showing ${obj}...`,
   counting: obj => `Counting ${obj}...`,
   depositing: () => 'Making deposit...',
+  checkingSwapRate: () => 'Checking swap rate...',
   claiming: obj => `Claiming ${obj}...`,
-  deposited: amount => `Deposited ${amount} nRLC to your iExec account`,
+  deposited: amount => `Deposited ${amount} RLC to your iExec account`,
   withdrawing: () => 'Making withdraw...',
-  withdrawn: amount => `${amount} nRLC withdrawn from your iExec account`,
+  withdrawn: amount => `${amount} RLC withdrawn from your iExec account`,
   downloading: () => 'Downloading result',
   decrypting: () => 'Decrypting result',
   downloaded: filePath => `Downloaded task result to file ${filePath}`,
@@ -58,13 +60,13 @@ const desc = {
   showObj: (objName, owner = 'user') => `show ${owner} ${objName} details`,
   countObj: (objName, owner = 'user') => `get ${owner} ${objName} count`,
   claimObj: objName => `claim a ${objName} that is not COMPLETED`,
-  deposit: () => 'deposit RLC onto your iExec account',
-  withdraw: () => 'withdraw RLC from your iExec account',
-  getETH: () => 'apply for ETH from pre-registered faucets',
-  getRLC: () => 'apply for nRLC from iExec faucet',
-  sendETH: () => 'send ETH to an address',
-  sendRLC: () => 'send nRLC to an address',
-  sweep: () => 'send all ETH and RLC to an address',
+  deposit: () => 'deposit RLC onto your iExec account (default unit nRLC)',
+  withdraw: () => 'withdraw RLC from your iExec account (default unit nRLC)',
+  getETH: () => 'apply for test ether from pre-registered faucets',
+  getRLC: () => 'apply for test RLC from iExec faucet',
+  sendETH: () => 'send ether to an address (default unit ether)',
+  sendRLC: () => 'send RLC to an address (default unit nRLC)',
+  sweep: () => 'send all ether and RLC to an address',
   info: () => 'show iExec contracts addresses',
   validateRessource: () => 'validate an app/dataset/workerpool description before submitting it to the iExec registry',
   decrypt: () => 'decrypt work result',
@@ -78,8 +80,8 @@ const desc = {
   encryptDataset: () => 'generate a key and encrypt the datasets files from "./datasets/original"',
   generateKeys: () => 'generate a beneficiary key pair to encrypt and decrypt the results',
   decryptResults: () => 'decrypt encrypted results with beneficary key',
-  bridgeToSidechain: () => 'send nRLC from the mainchain to the sidechain',
-  bridgeToMainchain: () => 'send nRLC from the sidechain to the mainchain',
+  bridgeToSidechain: () => 'send RLC from the mainchain to the sidechain (default unit nRLC)',
+  bridgeToMainchain: () => 'send RLC from the sidechain to the mainchain (default unit nRLC)',
   appRun: () => 'run an iExec application at market price (default run last deployed app)',
   initStorage: () => 'initialize the remote storage',
   checkStorage: () => 'check if the remote storage is initialized',
@@ -196,7 +198,7 @@ const option = {
   orderbookApp: () => ['--app <address>', 'filter by app address'],
   orderbookDataset: () => ['--dataset <address>', 'filter by dataset address'],
   requiredTag: () => [
-    '--require-tag <tag...>',
+    '--require-tag <tag>',
     'specify minimum required tags\n* usage: --require-tag tag1,tag2',
   ],
   password: () => [
@@ -258,8 +260,8 @@ const option = {
     ])}>`,
   ],
   txGasPrice: () => [
-    '--gas-price <wei>',
-    'set custom gas price for transactions (in wei)',
+    '--gas-price <amount unit...>',
+    'set custom gas price for transactions (default unit wei)',
   ],
   forceUpdateSecret: () => ['--force-update', 'update if already exists'],
   storageToken: () => [
@@ -285,18 +287,24 @@ const orderOption = {
     `--workerpool <${listOfChoices(['deployed'], 'address')}>`,
     'workerpool address, use "deployed" to use last deployed from "deployed.json"',
   ],
-  price: () => ['--price <nRlcAmount>', 'price per task'],
-  appprice: () => ['--app-price <nRlcAmount>', 'app price per task'],
+  price: () => [
+    '--price <amount unit...>',
+    'price per task (default unit nRLC)',
+  ],
+  appprice: () => [
+    '--app-price <amount unit...>',
+    'app price per task (default unit nRLC)',
+  ],
   datasetprice: () => [
-    '--dataset-price <nRlcAmount>',
-    'dataset price per task',
+    '--dataset-price <amount unit...>',
+    'dataset price per task (default unit nRLC)',
   ],
   workerpoolprice: () => [
-    '--workerpool-price <nRlcAmount>',
-    'workerpool price per task',
+    '--workerpool-price <amount unit...>',
+    'workerpool price per task (default unit nRLC)',
   ],
   volume: () => ['--volume <volume>', 'number of run'],
-  tag: () => ['--tag <tag...>', 'specify tags\n* usage: --tag tag1,tag2'],
+  tag: () => ['--tag <tag>', 'specify tags\n* usage: --tag tag1,tag2'],
   category: () => ['--category <id>', 'id of the task category'],
   trust: () => ['--trust <integer>', 'trust level'],
   apprestrict: () => [
@@ -332,7 +340,7 @@ const orderOption = {
     'specify the arguments to pass to the app',
   ],
   requestInputFiles: () => [
-    '--input-files <fileUrl...>',
+    '--input-files <fileUrl>',
     'specify the URL of input files to be used by the app\n* usage: --input-files https://example.com/foo.txt,https://example.com/bar.zip',
   ],
   requestEncryptResult: () => [
@@ -365,13 +373,14 @@ const addWalletLoadOptions = (cli) => {
 
 const question = async (
   message,
-  { error = info.userAborted(), strict = true } = {},
+  { error = info.userAborted(), rejectDefault = false, strict = true } = {},
 ) => {
   const answer = await inquirer.prompt([
     {
       type: 'confirm',
       name: 'ok',
       message,
+      default: !rejectDefault,
     },
   ]);
   if (answer.ok) return true;
@@ -422,11 +431,9 @@ const prompt = {
   transfer: (currency, amount, chainName, to, chainId) => question(
     `Do you want to send ${amount} ${chainName} ${currency} to ${to} [chainId: ${chainId}]`,
   ),
-  fillOrder: (amount, orderID) => question(
-    `Do you want to spend ${amount} nRLC to fill order with ID ${orderID} and submit your work`,
-  ),
-  placeOrder: (volume, category, value) => question(
-    `Do you want to place a sell order for ${volume} work category ${category} at ${value} nRLC each`,
+  swapEthToRlc: (etherAmount, rlcAmount) => question(
+    `Do you want to spend ${etherAmount} ether from your wallet to received ${rlcAmount} RLC on your account?`,
+    { rejectDefault: true },
   ),
   cancelOrder: (orderName, order) => question(`Do you want to cancel the following ${orderName}? ${order}`),
   publishOrder: (orderName, order) => question(`Do you want to publish the following ${orderName}? ${order}`),
@@ -441,14 +448,11 @@ const prompt = {
   limitedVolume: (available, ask) => question(
     `Your user order is valid for ${ask} work executions but other orders allow only ${available} work executions, do you want to continue?`,
   ),
-  limitedStake: (totalCost, stake, payableVolume) => question(
-    `total cost is ${totalCost} nRLC and you have ${stake} nRLC staked on your account. Your stake allows you to purchase ${payableVolume} work, do you want to continue?`,
-  ),
   unpublishFromJsonFile: (orderName, order) => question(`Do you want to unpublish the following ${orderName}? ${order}`),
 };
 
-prompt.transferETH = (...args) => prompt.transfer('ETH', ...args);
-prompt.transferRLC = (...args) => prompt.transfer('nRLC', ...args);
+prompt.transferETH = (...args) => prompt.transfer('ether', ...args);
+prompt.transferRLC = (...args) => prompt.transfer('RLC', ...args);
 prompt.sweep = currencies => (...args) => prompt.transfer(currencies, 'all wallet', ...args);
 
 const oraOptions = {
@@ -679,12 +683,13 @@ const DEFAULT_DECRYPTED_RESULTS_NAME = 'results.zip';
 const publicKeyName = address => `${address}_key.pub`;
 const privateKeyName = address => `${address}_key`;
 
-const computeTxOptions = (opts) => {
+const computeTxOptions = async (opts) => {
   let gasPrice;
   if (opts.gasPrice) {
-    if (!/^\d+$/i.test(opts.gasPrice)) throw Error('Invalid gas price value');
-    const bnGasPrice = new BN(opts.gasPrice);
-    if (bnGasPrice.isNeg()) throw Error('Invalid gas price, must be positive');
+    debug('opts.gasPrice', opts.gasPrice);
+    const bnGasPrice = new BN(
+      await weiAmountSchema({ defaultUnit: 'wei' }).validate(opts.gasPrice),
+    );
     gasPrice = '0x'.concat(bnGasPrice.toString('hex'));
   }
   debug('gasPrice', gasPrice);